Add audio support to the capture test.
Work-in-progress commit. Add a GIPSAudioSource object to
the capture demo, and call it to capture audio. It uses
the ExternalMediaProcessing interface to install
a callback, which is called for each audio buffer.

Eventually the callback would spool to disk, or a ringbuffer
with a separate thread spooling to disk if we can't do that
in 10 ms.

Note that no audio flows at all without setting up an RTP/RTCP
connection. That is, the video capture will still run, but
the Process() callback will never be invoked, and there's
to monitor signal sent to the playout device. It may be
possible to use the ExternalTransport mechanism to short-circuit
this.